rubocop適用に伴う修正

+++ app/controllers/orders_controller.rb
@@ -1,27 +1,18 @@
+# frozen_string_literal: true
+
 class OrdersController < ApplicationController
   def create
     @order = Order.new(order_params)
     @cart_items = current_cart.cart_items.includes(:product)
 
-    ApplicationRecord.transaction do
-      @order.save!
-      @cart_items.each do |cart_item|
-        product = cart_item.product
-        @order.order_details.create!(
-          product_name: product.name,
-          product_price: product.price,
-          quantity: cart_item.quantity,
-          subtotal: cart_item.subtotal
-        )
-      end
-      current_cart.destroy
+    if @order.save_with_details(current_cart)
       session.delete(:cart_id)
       OrderMailer.ordermail(@order).deliver_now
-      redirect_to root_url, notice: "購入ありがとうございます"
+      redirect_to root_url, notice: '購入ありがとうございます'
+    else
+      flash.now[:error] = '注文が正常に行えませんでした'
+      render 'carts/show', status: :unprocessable_entity
     end
-  rescue ActiveRecord::RecordInvalid
-    flash.now[:error] = "注文が正常に行えませんでした"
-    render "carts/show", status: :unprocessable_entity
   end
 
   private
